requirements.txt
alembic==1.5.6
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Keras==2.4.3
mypy==0.812
pandas==1.2.3
pytest==6.2.2
python-dotenv==0.15.0
SQLAlchemy==1.3.23
tensorflow==2.4.1
torch==1.8.0
wheel==0.36.2